@@ -6,7 +6,7 @@ import 'package:flutter/physics.dart';
 /// i extended  the functionality to be more useful in real world applications
 class StepperTouch extends StatefulWidget {
   const StepperTouch({
-    Key key,
+    Key? key,
     this.initialValue,
     this.onChanged,
     this.direction = Axis.horizontal,
@@ -20,10 +20,10 @@ class StepperTouch extends StatefulWidget {
   final Axis direction;
 
   /// the initial value of the stepper
-  final int initialValue;
+  final int? initialValue;
 
   /// called whenever the value of the stepper changed
-  final ValueChanged<int> onChanged;
+  final ValueChanged<int>? onChanged;
 
   /// if you want a springSimulation to happens the the user let go the stepper
   /// defaults to true
@@ -37,35 +37,31 @@ class StepperTouch extends StatefulWidget {
   _Stepper2State createState() => _Stepper2State();
 }
 
-class _Stepper2State extends State<StepperTouch>
-    with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation _animation;
-  int _value;
-  double _startAnimationPosX;
-  double _startAnimationPosY;
+class _Stepper2State extends State<StepperTouch> with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<Offset> _animation;
+  late int _value;
+  late double _startAnimationPosX;
+  late double _startAnimationPosY;
 
   @override
   void initState() {
     super.initState();
     _value = widget.initialValue ?? 0;
-    _controller =
-        AnimationController(vsync: this, lowerBound: -0.5, upperBound: 0.5);
+    _controller = AnimationController(vsync: this, lowerBound: -0.5, upperBound: 0.5);
     _controller.value = 0.0;
     _controller.addListener(() {});
 
     if (widget.direction == Axis.horizontal) {
-      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(1.5, 0.0))
-          .animate(_controller);
+      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(1.5, 0.0)).animate(_controller);
     } else {
-      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(0.0, 1.5))
-          .animate(_controller);
+      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(0.0, 1.5)).animate(_controller);
     }
   }
 
   @override
   void dispose() {
-    _controller?.dispose();
+    _controller.dispose();
     super.dispose();
   }
 
@@ -73,11 +69,9 @@ class _Stepper2State extends State<StepperTouch>
   void didUpdateWidget(oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.direction == Axis.horizontal) {
-      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(1.5, 0.0))
-          .animate(_controller);
+      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(1.5, 0.0)).animate(_controller);
     } else {
-      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(0.0, 1.5))
-          .animate(_controller);
+      _animation = Tween<Offset>(begin: Offset(0.0, 0.0), end: Offset(0.0, 1.5)).animate(_controller);
     }
   }
 
@@ -98,8 +92,7 @@ class _Stepper2State extends State<StepperTouch>
               Positioned(
                 left: widget.direction == Axis.horizontal ? 10.0 : null,
                 bottom: widget.direction == Axis.horizontal ? null : 10.0,
-                child:
-                    Icon(Icons.remove, size: 40.0, color: widget.buttonsColor),
+                child: Icon(Icons.remove, size: 40.0, color: widget.buttonsColor),
               ),
               Positioned(
                 right: widget.direction == Axis.horizontal ? 10.0 : null,
@@ -119,16 +112,13 @@ class _Stepper2State extends State<StepperTouch>
                     child: Center(
                       child: AnimatedSwitcher(
                         duration: const Duration(milliseconds: 500),
-                        transitionBuilder:
-                            (Widget child, Animation<double> animation) {
-                          return ScaleTransition(
-                              child: child, scale: animation);
+                        transitionBuilder: (Widget child, Animation<double> animation) {
+                          return ScaleTransition(child: child, scale: animation);
                         },
                         child: Text(
                           '$_value',
                           key: ValueKey<int>(_value),
-                          style: TextStyle(
-                              color: widget.counterColor, fontSize: 56.0),
+                          style: TextStyle(color: widget.counterColor, fontSize: 56.0),
                         ),
                       ),
                     ),
@@ -175,26 +165,22 @@ class _Stepper2State extends State<StepperTouch>
       changed = true;
     }
     if (widget.withSpring) {
-      final SpringDescription _kDefaultSpring =
-          new SpringDescription.withDampingRatio(
+      final SpringDescription _kDefaultSpring = new SpringDescription.withDampingRatio(
         mass: 0.9,
         stiffness: 250.0,
         ratio: 0.6,
       );
       if (widget.direction == Axis.horizontal) {
-        _controller.animateWith(
-            SpringSimulation(_kDefaultSpring, _startAnimationPosX, 0.0, 0.0));
+        _controller.animateWith(SpringSimulation(_kDefaultSpring, _startAnimationPosX, 0.0, 0.0));
       } else {
-        _controller.animateWith(
-            SpringSimulation(_kDefaultSpring, _startAnimationPosY, 0.0, 0.0));
+        _controller.animateWith(SpringSimulation(_kDefaultSpring, _startAnimationPosY, 0.0, 0.0));
       }
     } else {
-      _controller.animateTo(0.0,
-          curve: Curves.bounceOut, duration: Duration(milliseconds: 500));
+      _controller.animateTo(0.0, curve: Curves.bounceOut, duration: Duration(milliseconds: 500));
     }
 
     if (changed && widget.onChanged != null) {
-      widget.onChanged(_value);
+      widget.onChanged!(_value);
     }
   }
 }
